refactor: move address-book reconnect overlay into features/contacts
