+++ src/renderer/features/contacts/BackendContacts/index.ts
@@ -1,2 +1,4 @@
 export { backendContactsModel } from './model/backend-contacts-model';
 export { SyncStatusBadge } from './ui/SyncStatusBadge';
+export { AddressBookHealthOverlay } from './ui/AddressBookHealthOverlay';
+export { ReconnectAddressBookButton } from './ui/ReconnectAddressBookButton';
