+++ src/renderer/features/drafts/components/DraftsSection.tsx
@@ -20,7 +20,7 @@ import { contactModel } from '@/entities/contact';
 import { networkModel, useApi } from '@/entities/network';
 import { accountUtils, walletModel, walletUtils } from '@/entities/wallet';
 import { authModel, backendConfigurationModel } from '@/aggregates/backend';
-import { backendContactsModel } from '@/features/contacts';
+import { AddressBookHealthOverlay, backendContactsModel } from '@/features/contacts';
 import { tryDecodeCallData } from '../lib/decode-call-data';
 import { resolveDraftProxyAccount } from '../lib/draft-account-resolution';
 import { useCanCreateDraft } from '../lib/useCanCreateDraft';
@@ -31,7 +31,6 @@ import { draftDeepLinkModel } from '../model/draft-deep-link';
 import '../model/drafts-model'; // side-effect: orchestration wiring
 import { submitDraftModel } from '../model/submit-draft-model';
 
-import { AddressBookHealthOverlay } from './AddressBookHealthOverlay';
 import { DraftRow } from './DraftRow';
 import { DraftSummary } from './DraftSummary';
